Reorganization and first multi publications test
