+++ src/test/static_publications/__init__.py
