+++ src/test/static_publications/single_publication/maps_test.py
@@ -3,7 +3,7 @@ import pytest
 
 from layman import app, util as layman_util
 from test_tools import process_client
-from ... import single_static_publication as data
+from ... import static_publications as data
 from ..data import ensure_publication
