+++ src/test/static_publications/single_publication/publications_test.py
@@ -4,7 +4,7 @@ from layman.common.prime_db_schema import publications
 from layman.layer.filesystem import gdal, thumbnail as layer_thumbnail
 from layman.map.filesystem import thumbnail as map_thumbnail
 from test_tools import assert_util, util as test_util, process_client
-from ... import single_static_publication as data
+from ... import static_publications as data
 from ..data import ensure_publication
 
 
@@ -49,23 +49,6 @@ def test_user_workspace(workspace, publ_type, publication):
     else:
         assert workspace not in usernames
     assert workspace in workspaces
-
-
-@pytest.mark.parametrize('context, expected_publications', [
-    ({'actor_name': 'test_get_publication_infos_user_actor', 'access_type': 'read'}, {'post_public_sld', 'post_private_write_sld'},),
-    ({'actor_name': 'test_get_publication_infos_user_actor', 'access_type': 'write'}, {'post_public_sld'},),
-], )
-@pytest.mark.usefixtures('liferay_mock', 'ensure_layman',)
-def test_get_publication_infos(context,
-                               expected_publications):
-    ensure_publication(data.OWNER, process_client.LAYER_TYPE, 'post_private_sld')
-    ensure_publication(data.OWNER, process_client.LAYER_TYPE, 'post_private_write_sld')
-    ensure_publication(data.OWNER, process_client.LAYER_TYPE, 'post_public_sld')
-
-    with app.app_context():
-        infos = layman_util.get_publication_infos(data.OWNER, process_client.LAYER_TYPE, context)
-    publ_set = set(publication_name for (workspace, publication_type, publication_name) in infos.keys())
-    assert expected_publications.issubset(publ_set), publ_set
 
 
 @pytest.mark.parametrize('workspace, publ_type, publication', data.LIST_ALL_PUBLICATIONS)
